lib/oauth: remove the Complete() method from the interface, bring consistency.
Background:
astore server checks that authentication was completed successuffully
using the moral equivalent of 'authenticator.Complete(receivedAuthData)'.

The oauth library itself, in the provided/ready-to-use auth handlers...
does not. It just checks that some parameters are not nil.

authenticator.Complete() does some of the nil checks, plus a couple
more for defense in depth. One of those checks is ... it reparses
the cookie, that PerformAuth just parsed and returned in the auth!

In order to prepare for the next PRs, turning some methods in free
standing functions and improving some of the interfaces, this PR:
- simplifies Complete() by removing the re-parsing of the cookie.
  (the only way to get an AuthData() struct is via PerformAuth that... parsed
  the cookie - if a cookie is used by the authentication method)
- moves Complete() from the authenticator object into the AuthData
  itself - wether AuthData is valid or not is only a function of
  the content of AuthData now.
- removes the need for IAuthenticator implementations to provide
  a Complete() method.

(last point was pretty much the goal of the PR)