lib/oauth: clarify the IAuthenticator interface, implement it in redirector.
Background:
lib/oauth evolved over time to be more of an authentication library
than just an oauth library. Eg, at the end of the oauth process (or
through any other authentication process), it assigns and manages
encrypted cookies carrying principals and groups, and provides
primitives to use and validate those cookies.

It currently provides two ways to authenticate users:
- via oauth - redirecting to an IdP and using the full oauth protocol.
- via simple redirect to an internal server - which then sets cookies
  that the backend can interpret.

In this PR:
- improved the documentation to better represent the evolution
  of the library.
- make the Redirector authenticator implement the IAuthenticator
  interface, by refactoring the Authenticate() method.
- a couple other changes to bring consistency across the different
  implementations of IAuthenticator - mostly around error values
  returned and validation.