lib/oauth: propagate target url in redirector
Background:
The "Redirector" in oauth lib authenticates a user by redirecting
him/her to an "auth server", and then checking a cookie.
At the end of the auth, the auth server needs to redirect the user
back to the page he/she was on - the end "target URL".

Before this PR:
- users of the API always had to explicitly specify the
  "target URL". This is in contrast with the oauth authenticator,
  that has by default a target url the user is sent back to if
  no specific url is configured via APIs.

In this PR:
- propagate the default target url to the "Redirector" object.
- use the default target url if one url was not explicitly
  specified in the call to the login API.