enkit/client: add --console-format=pretty-json to astore commands. (#1266)
Background:
All astore commands support the `--console-format=...` flag, which
controls the output format.
For example, `enkit astore list --console-format=json` will output JSON
instead of the usual tabular format.

This commit:
* adds the --console-format=pretty-json, same as --console-format=json,
but
  with indentation and newlines, better consumable by humans.

This is handy as it shows information that would otherwise not be
shown in the tabular format (for example, the sid).