lib/khttp/kassets: adding helper functions to handle "new style" go embeds.
Background:
When this repository was created, it was common to embed data in go binaries
using `go_embed_data` rules. This has been deprecated upstream in favor of
attributes in the normal go_library/go_binary rules.

The new mechanism built on the native golang support for embed uses an
embed.FS, while the old mechanism used a simple dict.

In this PR:
- added a couple functions to convert embeds to the old style dict,
  still used by a lot of functions within this repository.

This allows new code to use the native golang embed support with
the libraries, providing a path to deprecating go_embed_data.